mysite/requirements.txt
asgiref==3.8.1
astroid==3.1.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
defusedxml==0.7.1
dill==0.3.8
dj-rest-auth==5.0.2
Django==5.0.3
django-allauth==0.61.1
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
docopt==0.6.2
graphviz==0.20.3
idna==3.6
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.3
Js2Py==0.74
MarkupSafe==2.1.5
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
pipwin==0.5.2
platformdirs==4.2.0
psycopg2-binary==2.9.9
pycparser==2.21
pydotplus==2.0.2
pyjsparser==2.7.1
PyJWT==2.8.0
pylint==3.1.0
pyparsing==3.1.2
PyPrind==2.11.3
pySmartDL==1.3.4
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tomlkit==0.12.4
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
